Update to injury data
Added function to convert all raw pdf reports to csv files that can be used to generate list of players to not use in optimization

+++ src/injury_data.py
@@ -3,10 +3,13 @@ import re
 import datetime
 import pandas as pd
 import os
+import pdfreader
+from pdfreader import SimplePDFViewer
+from tabula import read_pdf, convert_into
 
 today = datetime.datetime.today()
 today = today.strftime('%Y-%m-%d')
-
+ROOT_DIR = os.path.abspath(os.curdir)
 
 def download_nba_injuryreports(date):
     # given a date in %Y-%m-%d string format, will try to download the injury reports that the nba publishes at 11, 2, and 5
@@ -17,13 +20,28 @@ def download_nba_injuryreports(date):
     r2 = requests.get(url2, allow_redirects=True)
     r3 = requests.get(url3, allow_redirects=True)
 
-    os.chdir('C:\\Users\\Kevin\\Documents\\GitHub\\draftkings\\data\\raw\\injury_reports')
-    if r1.status_code == 200:
+    path = ROOT_DIR + '\\data\\raw\\injury_reports'
+
+    if (r1.status_code == 200) and (os.path.exists(path + '\\' + 'Injury-Report_' + date + '_11AM.pdf') is False):
         open('Injury-Report_' + date + '_11AM.pdf', 'wb').write(r1.content)
-    if r2.status_code == 200:
+    if (r2.status_code == 200) and (os.path.exists(path + '\\' + 'Injury-Report_' + date + '_02PM.pdf') is False):
         open('Injury-Report_' + date + '_02PM.pdf', 'wb').write(r2.content)
-    if r3.status_code == 200:
+    if (r3.status_code == 200) and (os.path.exists(path + '\\' + 'Injury-Report_' + date + '_05PM.pdf') is False):
         open('Injury-Report_' + date + '_05PM.pdf', 'wb').write(r3.content)
 
 # download_nba_injuryreports(today)
+
+# converts pdf injury reports into csv files
+def parse_injuryreports():
+    path = ROOT_DIR + '\\data\\raw\\injury_reports'
+    for file in os.listdir(path):
+        df = read_pdf(
+            path + "\\" + file,
+            pages='all', stream=True, area=[49.99375, 13.15625, 563.61375, 832.00125])
+        final_report = pd.DataFrame()
+        for i in range(0, len(df)):
+            page = pd.DataFrame(df[i])
+            final_report = final_report.append(page)
+        if(os.path.exists(ROOT_DIR + '\\data\\processed\\injury_reports' + "\\" + file[:-3] + "csv") is False):
+            final_report.to_csv(ROOT_DIR + '\\data\\processed\\injury_reports' + "\\" + file[:-3] + "csv")
